refactor(mobile): prefetch entry logic
