+++ apps/mobile/src/modules/entry-list/entry-list-gird.tsx
@@ -6,14 +6,15 @@ import type { MasonryFlashListProps } from "@shopify/flash-list"
 import { MasonryFlashList } from "@shopify/flash-list"
 import { Image } from "expo-image"
 import { Link } from "expo-router"
-import { useContext, useState } from "react"
+import { useContext } from "react"
 import { Pressable, View } from "react-native"
 import { useSafeAreaInsets } from "react-native-safe-area-context"
 
 import { NavigationContext } from "@/src/components/common/SafeNavigationScrollView"
 import { ThemedText } from "@/src/components/common/ThemedText"
 import { ItemPressable } from "@/src/components/ui/pressable/item-pressable"
-import { useEntry, useFetchEntryContentByStream } from "@/src/store/entry/hooks"
+import { useEntry } from "@/src/store/entry/hooks"
+import { debouncedFetchEntryContentByStream } from "@/src/store/entry/store"
 
 import { useSelectedFeed } from "../feed-drawer/atoms"
 
@@ -23,9 +24,6 @@ export function EntryListContentGrid({
 }: {
   entryIds: string[]
 } & Omit<MasonryFlashListProps<string>, "data" | "renderItem">) {
-  const [viewableEntryIds, setViewableEntryIds] = useState<string[]>([])
-  useFetchEntryContentByStream(viewableEntryIds)
-
   const insets = useSafeAreaInsets()
   const tabBarHeight = useBottomTabBarHeight()
   const headerHeight = useHeaderHeight()
@@ -38,7 +36,7 @@ export function EntryListContentGrid({
       }, [])}
       keyExtractor={(id) => id}
       onViewableItemsChanged={({ viewableItems }) => {
-        setViewableEntryIds(viewableItems.map((item) => item.key))
+        debouncedFetchEntryContentByStream(viewableItems.map((item) => item.key))
       }}
       numColumns={2}
       onScroll={useTypeScriptHappyCallback(
