+++ apps/mobile/src/store/entry/store.ts
@@ -1,6 +1,9 @@
 import { FeedViewType } from "@follow/constants"
+import { debounce } from "es-toolkit/compat"
+import { fetch as expoFetch } from "expo/fetch"
 
 import { apiClient } from "@/src/lib/api-fetch"
+import { getCookie } from "@/src/lib/auth"
 import { honoMorph } from "@/src/morph/hono"
 import { storeDbMorph } from "@/src/morph/store-db"
 import { EntryService } from "@/src/services/entry"
@@ -215,7 +218,86 @@ class EntrySyncServices {
     }
     return entry
   }
+
+  async fetchEntryContentByStream(remoteEntryIds?: string[]) {
+    if (!remoteEntryIds || remoteEntryIds.length === 0) return
+
+    const onlyNoStored = true
+
+    const nextIds = [] as string[]
+    if (onlyNoStored) {
+      for (const id of remoteEntryIds) {
+        const entry = getEntry(id)!
+        if (entry.content) {
+          continue
+        }
+
+        nextIds.push(id)
+      }
+    }
+
+    if (nextIds.length === 0) return
+
+    const readStream = async () => {
+      // https://github.com/facebook/react-native/issues/37505
+      // TODO: And it seems we can not just use fetch from expo for ofetch, need further investigation
+      const response = await expoFetch(apiClient.entries.stream.$url().toString(), {
+        method: "post",
+        headers: {
+          cookie: getCookie(),
+        },
+        body: JSON.stringify({
+          ids: nextIds,
+        }),
+      })
+
+      const reader = response.body?.getReader()
+      if (!reader) return
+
+      const decoder = new TextDecoder()
+      let buffer = ""
+
+      try {
+        while (true) {
+          const { done, value } = await reader.read()
+          if (done) break
+
+          buffer += decoder.decode(value, { stream: true })
+          const lines = buffer.split("\n")
+
+          // Process all complete lines
+          for (let i = 0; i < lines.length - 1; i++) {
+            if (lines[i]!.trim()) {
+              const json = JSON.parse(lines[i]!)
+              // Handle each JSON line here
+              entryActions.updateEntryContent(json.id, json.content)
+            }
+          }
+
+          // Keep the last incomplete line in the buffer
+          buffer = lines.at(-1) || ""
+        }
+
+        // Process any remaining data
+        if (buffer.trim()) {
+          const json = JSON.parse(buffer)
+
+          entryActions.updateEntryContent(json.id, json.content)
+        }
+      } catch (error) {
+        console.error("Error reading stream:", error)
+      } finally {
+        reader.releaseLock()
+      }
+    }
+
+    readStream()
+  }
 }
 
 export const entrySyncServices = new EntrySyncServices()
 export const entryActions = new EntryActions()
+export const debouncedFetchEntryContentByStream = debounce(
+  entrySyncServices.fetchEntryContentByStream,
+  1000,
+)
